+++ actions/marketplace.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import ABI from "@/ABIs/marketplaceAbi.json";
-import { createContract } from "./actions";
+import { createContract } from "./serverActions";
 const contractAddress = "0xB594Cdeb1b46254A11Fc69d25D1a726aEbf9642c";
 
 export async function getAllListing() {
