+++ components/Col.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 import Link from "next/link";
 import Attest from "./Attest2";
 import { useEffect, useState } from "react";
-import { isOwnerOf } from "@/actions/actions";
+import { isOwnerOf } from "@/actions/clientActions";
 import Purchase from "./Purchase";
 import { NFTData } from "@/redux/types";
 import { RootState } from "@/redux/store";
