+++ components/ConnectWallet.tsx
@@ -8,7 +8,7 @@ import { FaWallet } from "react-icons/fa";
 import { useDispatch } from "react-redux";
 import { setisConnected } from "@/redux/slices/isconnected.slice";
 import { setAccount } from "@/redux/slices/account.slice";
-import { getProvider } from "@/actions/actions";
+import { getProvider } from "@/actions/clientActions";
 
 interface ConnectWalletProps {
   className?: string;
